refresh process for testing
